requirements.txt
amqp==2.6.1
arabic-reshaper==2.1.1
asgiref==3.3.1
autopep8==1.5.5
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
cairocffi==1.0.0
CairoSVG==2.5.2
celery==4.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
cssselect2==0.4.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.6
django-admin-list-filter-dropdown==1.0.3
django-admin-thumbnails==0.2.5
django-allauth==0.44.0
django-ckeditor==6.0.0
django-crispy-forms==1.11.1
django-currencies==0.10.1
django-extensions==2.2.5
django-import-export==2.5.0
django-jalali==4.1.2
django-js-asset==1.2.2
django-modeltranslation==0.16.1
django-mptt==0.11.0
django-paypal==1.0.0
django-rosetta==0.9.5
et-xmlfile==1.0.1
future==0.18.2
html5lib==1.1
idna==2.10
jdatetime==3.6.2
jdcal==1.4.1
jsonfield==2.1.1
kombu==4.6.11
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
persiantools==2.0.0
Pillow==8.1.0
polib==1.1.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==19.0.0
PyPDF2==1.26.0
Pyphen==0.10.0
python-bidi==0.4.2
python-gettext==4.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
reportlab==3.5.66
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==3.1.0
social-auth-core==4.1.0
soupsieve==2.2
sqlparse==0.4.1
tablib==3.0.0
tinycss2==1.1.0
toml==0.10.2
urllib3==1.26.3
vine==1.3.0
webencodings==0.5.1
Werkzeug==0.16.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
